requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.0
amqp==5.3.1
andi==0.8.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.1
apache-airflow-core==3.1.1
apache-airflow-providers-celery==3.12.4
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-http==5.3.4
apache-airflow-providers-postgres==6.3.0
apache-airflow-providers-redis==4.3.1
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.0
apache-airflow-task-sdk==1.1.1
argcomplete==3.6.3
asgiref==3.10.0
async-lru==2.0.5
asyncpg==0.30.0
attrs==25.4.0
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.14.2
billiard==4.2.2
black==25.9.0
cadwyn==5.4.5
celery==5.5.3
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.10.1
constantly==23.10.4
coverage==7.11.0
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.0
cssselect==1.3.0
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
environ==1.0
Faker==37.12.0
fastapi==0.117.1
fastapi-cli==0.0.14
filelock==3.20.0
flower==2.0.1
frozenlist==1.8.0
fsspec==2025.9.0
googleapis-common-protos==1.71.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.14.0
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.3.0
itemadapter==0.12.2
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
# Editable install with no version control (job_finder==0.1.0)
-e /home/harrison/Workspace/pessoal/JobFinder
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
lazy-object-proxy==1.12.0
libcst==1.8.5
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
natsort==8.4.0
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outcome==1.3.0.post0
packaging==25.0
parsel==1.10.0
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
Protego==0.5.0
protobuf==6.33.0
psutil==7.1.1
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyDispatcher==2.0.7
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
PyYAML-ft==8.0.0
queuelib==1.8.0
redis==5.2.1
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
retryhttp==1.3.3
rich==14.2.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.28.0
ruff==0.14.2
Scrapy==2.13.3
scrapy-poet==0.26.0
service-identity==24.2.0
setproctitle==1.3.7
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlitedict==2.1.0
sqlparse==0.5.3
starlette==0.48.0
statsd==4.0.1
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
time-machine==2.19.0
tldextract==5.3.0
tornado==6.5.2
Twisted==25.5.0
typer==0.20.0
types-python-dateutil==2.9.0.20251008
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
url-matcher==0.6.0
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
w3lib==2.3.1
watchfiles==1.1.1
wcwidth==0.2.14
web-poet==0.20.0
websockets==15.0.1
wheel==0.45.1
wirerope==1.0.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zope.interface==8.0.1
